+++ src/subscript-loc.c
@@ -182,7 +182,10 @@ static void int_check_consecutive(SEXP subscript, R_len_t n, R_len_t n_extend,
   }
 
   if (n_extend != i_extend) {
-    Rf_error("Internal error in `int_check_consecutive()`: n_extend != i_extend.");
+    stop_internal("int_check_consecutive",
+                  "n_extend (%d) != i_extend (%d).",
+                  n_extend,
+                  i_extend);
   }
 
   if (i_extend == 0) {
@@ -326,8 +329,7 @@ SEXP vec_as_location_opts(SEXP subscript, R_len_t n, SEXP names,
   case REALSXP: out = dbl_as_location(subscript, n, opts); break;
   case LGLSXP: out = lgl_as_location(subscript, n, opts); break;
   case STRSXP: out = chr_as_location(subscript, names, opts); break;
-  default: Rf_errorcall(R_NilValue, "Internal error: Wrong subscript type `%s` in `vec_as_location_opts()`.",
-                        Rf_type2char(TYPEOF(subscript)));
+  default: stop_unimplemented_type("vec_as_location_opts", TYPEOF(subscript));
   }
 
   UNPROTECT(2);
@@ -417,7 +419,7 @@ SEXP vctrs_as_location(SEXP subscript, SEXP n_, SEXP names,
     PROTECT(n_);
 
     if (Rf_length(n_) != 1) {
-      Rf_error("Internal error: `n` must be a scalar number");
+      stop_internal("vctrs_as_location", "`n` must be a scalar number.");
     }
 
     n = r_int_get(n_, 0);
@@ -442,8 +444,7 @@ SEXP vctrs_as_location(SEXP subscript, SEXP n_, SEXP names,
 
 static void stop_subscript_missing(SEXP i) {
   vctrs_eval_mask1(Rf_install("stop_subscript_missing"),
-                   syms_i, i,
-                   vctrs_ns_env);
+                   syms_i, i);
   never_reached("stop_subscript_missing");
 }
 
@@ -453,8 +454,7 @@ static void stop_location_negative_missing(SEXP i,
   vctrs_eval_mask3(Rf_install("stop_location_negative_missing"),
                    syms_i, i,
                    syms_subscript_arg, arg,
-                   syms_subscript_action, get_opts_action(opts->subscript_opts),
-                   vctrs_ns_env);
+                   syms_subscript_action, get_opts_action(opts->subscript_opts));
   never_reached("stop_location_negative_missing");
 }
 static void stop_location_negative_positive(SEXP i,
@@ -463,8 +463,7 @@ static void stop_location_negative_positive(SEXP i,
   vctrs_eval_mask3(Rf_install("stop_location_negative_positive"),
                    syms_i, i,
                    syms_subscript_arg, arg,
-                   syms_subscript_action, get_opts_action(opts->subscript_opts),
-                   vctrs_ns_env);
+                   syms_subscript_action, get_opts_action(opts->subscript_opts));
   never_reached("stop_location_negative_positive");
 }
 
@@ -477,8 +476,7 @@ static void stop_subscript_oob_location(SEXP i, R_len_t size,
                    syms_subscript_type, chrs_numeric,
                    syms_size, size_obj,
                    syms_subscript_action, get_opts_action(opts->subscript_opts),
-                   syms_subscript_arg, arg,
-                   vctrs_ns_env);
+                   syms_subscript_arg, arg);
 
   UNPROTECT(1);
   never_reached("stop_subscript_oob_location");
@@ -491,8 +489,7 @@ static void stop_subscript_oob_name(SEXP i, SEXP names,
                    syms_subscript_type, chrs_character,
                    syms_names, names,
                    syms_subscript_action, get_opts_action(opts->subscript_opts),
-                   syms_subscript_arg, arg,
-                   vctrs_ns_env);
+                   syms_subscript_arg, arg);
   never_reached("stop_subscript_oob_name");
 }
 
@@ -502,8 +499,7 @@ static void stop_location_negative(SEXP i,
   vctrs_eval_mask3(Rf_install("stop_location_negative"),
                    syms_i, i,
                    syms_subscript_action, get_opts_action(opts->subscript_opts),
-                   syms_subscript_arg, arg,
-                   vctrs_ns_env);
+                   syms_subscript_arg, arg);
   never_reached("stop_location_negative");
 }
 
@@ -513,8 +509,7 @@ static void stop_location_zero(SEXP i,
   vctrs_eval_mask3(Rf_install("stop_location_zero"),
                    syms_i, i,
                    syms_subscript_action, get_opts_action(opts->subscript_opts),
-                   syms_subscript_arg, arg,
-                   vctrs_ns_env);
+                   syms_subscript_arg, arg);
   never_reached("stop_location_zero");
 }
 
@@ -525,8 +520,7 @@ static void stop_indicator_size(SEXP i, SEXP n,
                    syms_i, i,
                    syms_n, n,
                    syms_subscript_action, get_opts_action(opts->subscript_opts),
-                   syms_subscript_arg, arg,
-                   vctrs_ns_env);
+                   syms_subscript_arg, arg);
   never_reached("stop_indicator_size");
 }
 
@@ -538,8 +532,7 @@ static void stop_location_oob_non_consecutive(SEXP i, R_len_t size,
                    syms_i, i,
                    syms_size, size_obj,
                    syms_subscript_action, get_opts_action(opts->subscript_opts),
-                   syms_subscript_arg, arg,
-                   vctrs_ns_env);
+                   syms_subscript_arg, arg);
 
   UNPROTECT(1);
   never_reached("stop_location_oob_non_consecutive");