+++ src/type-data-frame.c
@@ -81,7 +81,7 @@ SEXP vctrs_new_data_frame(SEXP args) {
 
     // We might add dynamic dots later on
     if (tag == R_ClassSymbol) {
-      Rf_error("Internal error in `new_data_frame()`: Can't supply `class` in `...`.");
+      stop_internal("new_data_frame", "Can't supply `class` in `...`.");
     }
 
     if (tag == R_NamesSymbol) {
@@ -171,23 +171,7 @@ static SEXP c_data_frame_class(SEXP cls) {
   if (TYPEOF(cls) != STRSXP) {
     Rf_errorcall(R_NilValue, "`class` must be NULL or a character vector");
   }
-  if (Rf_length(cls) == 0) {
-    return classes_data_frame;
-  }
-
-  SEXP args = PROTECT(Rf_allocVector(VECSXP, 2));
-  SET_VECTOR_ELT(args, 0, cls);
-  SET_VECTOR_ELT(args, 1, classes_data_frame);
-
-  SEXP out = vec_c(
-    args,
-    vctrs_shared_empty_chr,
-    R_NilValue,
-    NULL
-  );
-
-  UNPROTECT(1);
-  return out;
+  return chr_c(cls, classes_data_frame);
 }
 
 // [[ include("type-data-frame.h") ]]
@@ -486,7 +470,7 @@ SEXP df_cast_opts(const struct cast_opts* opts) {
   SEXP to_names = PROTECT(r_names(opts->to));
 
   if (x_names == R_NilValue || to_names == R_NilValue) {
-    Rf_error("Internal error in `df_cast()`: Data frame must have names.");
+    stop_internal("df_cast_opts", "Data frame must have names.");
   }
 
   SEXP out = R_NilValue;